fix(git): stabilize watcher reconciliation
